Unify widgets tests
